+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+### BREAKING CHANGES
+
+- Change URL of backend endpoint for getting the data of a single facsimile collection from `/<project>/facsimiles/collections/<collectionId>` to `/<project>/facsimile_collection/<collectionId>/data/`.
+
 
 
 ## [1.5.2] – 2025-11-27